+++ test/serializer/json/serializer_test.dart
@@ -284,11 +284,17 @@ void main() {
               '[16,239714735,{"exclude_me":false},"com.myapp.mytopic1",["Hello, world!"],{"color":"orange","sizes":[23,42,7]}]'));
     });
     test('Goodbye', () {
-      expect(serializer.serializeToString(Goodbye(GoodbyeMessage("cya"),Goodbye.REASON_GOODBYE_AND_OUT)),
+      expect(
+          serializer.serializeToString(
+              Goodbye(GoodbyeMessage("cya"), Goodbye.REASON_GOODBYE_AND_OUT)),
           equals('[6,{"message":"cya"},"wamp.error.goodbye_and_out"]'));
-      expect(serializer.serializeToString(Goodbye(GoodbyeMessage(null),Goodbye.REASON_CLOSE_REALM)),
+      expect(
+          serializer.serializeToString(
+              Goodbye(GoodbyeMessage(null), Goodbye.REASON_CLOSE_REALM)),
           equals('[6,{"message":""},"wamp.error.close_realm"]'));
-      expect(serializer.serializeToString(Goodbye(null,Goodbye.REASON_SYSTEM_SHUTDOWN)),
+      expect(
+          serializer
+              .serializeToString(Goodbye(null, Goodbye.REASON_SYSTEM_SHUTDOWN)),
           equals('[6,{},"wamp.error.system_shutdown"]'));
     });
   });